Separate functions of main.cpp into run.h
Make separate main function for testing
