+++ .gitignore
@@ -34,6 +34,7 @@
 # Output files
 output
 VEQ_S
+test_run
 
 # Other
 .DS_Store
